Feature: add read-current-options
Read-current-options obtains information from Yahoo regarding options. It diverges from the other API calls in that it returns nested hash tables of information for each requested symbol, instead of a a-list. I deemed the a-list to be too denormalized to start off with.

I might add a keyword to the other read-* functions that selects between an alist and a hash table.